requirements.txt
customtkinter==5.2.2
pillow==10.2.0
rcon @ git+https://github.com/conqp/rcon.git@5b2cb6c5a4304fa2e572460c6f83ada02d325317
